Fix ajax call failures in node.js
Ajax calls (of the type that would be made in nickleby-search)
failed with an error of 'No transport.'

This seems to result from a problem with jQuery.ajax in the
node environment (see http://stackoverflow.com/a/8916217).

This commit introduces a new devDependency, xmlhttprequest, and uses
an XHR from this module to replace the jQuery XHR function.

Incidentally, installing asn1.js, a dependency of xmlhttprequest, failed
on node v0.10.32. This commit installs successfully on node v0.10.22.
Future development will target this version (as enforced by the
.nvmrc file).